Set up aggregate data helper class
